add PhotosSetting, getPhoto(@Service)

+++ src/main/java/com/squarecross/photoalbum/domain/Photo.java
@@ -36,7 +36,7 @@ public class Photo {
     private String originalUrl;
 
     @Column(name = "file_size", unique = false, nullable = true)
-    private Long fileSize;
+    private int fileSize;
 
     @Column(name = "uploaded_at", unique = false, nullable = true)
     @CreationTimestamp
